powernv/TM: emulate tbegin on P10 as an illegal
P10 has an issue with tbegin which may cause the system to xstop hence
tbegin will be trapped. This is done via a softpatch interrupt
generated from the IMC.

The softpatch interrupt will go to 0x1500 with hsrr0 pointing to the
instruction after tbegin. For bare metal systems, TM is always off but
the softpatch take precedence over the facility unavailable, hence we
need to emulate the illegal instruction in the softpatch handler.

This patch does this.

FIXME: Another patch will be required for KVM. for P10 guests we do
the same and emulate an illegal instruction. For P8 and P9 guests, we
need to properly emulate the tbegin as a transaction failure (ie set
the CR bit, TEXASR bits, TFHAR, TFIAR etc)

Signed-off-by: Michael Neuling <mikey@neuling.org>